Multiple signatures on a document
This allows origami to sign the same document multiple times. In order
to do this, a revision needs added to the document. However the way
Origami works out of the gate is that the doc is parsed and then
rewritten causing some small changes that invalidates the exactness of
an e-signature.

I pursued a few ways to fix this but for accuracy and swiftness decided
it would be easiest to not modify the original document but to write the
original documents contents directly to the new file and then utilize
origami to generate the new revision which could be appended at the end.

This works fairly well but does have some issues with differences in bit
positions required for digital signatures. Thus there is currently quite
a bit of forking logic when the document is signed vs when it isn't. In
theory, this should be able to get unified to the same branch but it
takes quite a bit of time to test and validate the document so I haven't
gotten that far.

Lastly there are 2 methods. The normal sign method and the
sign_with_hsm method. I am not sure if the normal sign method works
properly. I utilized that up until I was happy with the direction /
feasibility and then switched to the sign_with_hsm method.

As you might have guessed, sign_with_hsm utilized a Hardare Security
Module (HSM) to store the key. In my deployment scenario I do not have
access to the private key, thus could not utilize OpenSSL on my machine
to do the signing.

Therefore there are some classes in the code that live in my application that
reimplements what OpenSSL does under the hood in generating the PKCS7
package and signing the generated package. I added them to this so you
could see where you need to reimplement that functionality. I can help
if you get in contact with me.